refactored sort logic
