+++ src/feature/sortSlice.js
@@ -13,38 +13,17 @@ const sortSlice = createSlice({
     // saving the use input
     saveInput: (state, action) => {
       state.inputArr = action.payload;
-      // state.inputArr = [...state.inputArr, action.payload];
     },
 
-    // sorting the user input array
-    sortArr: (state) => {
-      const inputArrCopy = [...state.inputArr];
-      let sortSteps = [];
-      let compareIndices = [];
-
-      // sorting logic
-      for (let i = 0; i < inputArrCopy.length; i++) {
-        for (let j = 1; j < inputArrCopy.length - i; j++) {
-          if (inputArrCopy[j] < inputArrCopy[j - 1]) {
-            let temp = inputArrCopy[j - 1];
-            inputArrCopy[j - 1] = inputArrCopy[j];
-            inputArrCopy[j] = temp;
-          }
-
-          // visualization logic
-          sortSteps.push([...inputArrCopy]);
-          compareIndices.push([j - 1, j]);
-        }
-      }
-      // sorted array
+    // saving the sorting result and the sorting steps for the visualization
+    setSortData: (state, action) => {
+      const { inputArrCopy, sortSteps, compareIndices } = action.payload;
       state.inputArr = inputArrCopy;
-
-      // visualization logic
       state.sortSteps = sortSteps;
       state.compareIndices = compareIndices;
     },
   },
 });
 
-export const { saveInput, sortArr } = sortSlice.actions;
+export const { saveInput, sortArr, setSortData } = sortSlice.actions;
 export default sortSlice.reducer;
